requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
Werkzeug==2.1.0
